Redefine schema for widgets and highcharts
Highcharts schema defined earlier with embedded schemas did
not define the structs properly enough to be used for the
desired functionality. It had to be redefined with a struct
to provide enough flexibility to be used in the application.

The widget schema was also refactored to be based on generic keys
defined by the application, so it can cater to widgets of different
types and from different vendors.

All the model files were also moved under widget directory to
envelope all the files in a single directory for a singel app.